+++ src/main/clojure/com/github/clojure_lsp/intellij/server.clj
@@ -5,6 +5,8 @@
    [com.github.clojure-lsp.intellij.db :as db]
    [com.github.clojure-lsp.intellij.logger :as logger]
    [com.github.clojure-lsp.intellij.client :as lsp-client]
+   [com.github.clojure-lsp.intellij.notification]
+   [com.github.clojure-lsp.intellij.workspace-edit]
    [com.github.clojure-lsp.intellij.tasks :as tasks]
    [lsp4clj.server :as lsp4clj.server])
   (:import